uploader/requirements.txt
kafka-python==1.3.4
tqdm
